# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Drammers.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/news/{id}": {
      "get": {
        "tags": [
          "AdminNews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminNewsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminNewsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminNewsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminNews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "AdminNews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/news": {
      "get": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfNewsSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfNewsSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfNewsSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/{id}": {
      "get": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminNewsResponse": {
        "required": [
          "id",
          "title",
          "summary",
          "body",
          "category",
          "expireAt",
          "publication",
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "body": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "expireAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationResponse"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ContentVisibility": {
        "enum": [
          "Public",
          "Members",
          "Restricted"
        ],
        "description": "Wie content mag zien (docs/07 §4). Beheren gaat via permissions, niet via zichtbaarheid."
      },
...
      "NewsDetailResponse": {
        "required": [
          "id",
          "title",
          "summary",
          "bodyHtml",
          "category",
          "publishedAt",
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "bodyHtml": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "NewsRequest": {
        "required": [
          "title",
          "summary",
          "body",
          "category",
          "expireAt",
          "publication"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "body": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "expireAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationRequest"
          }
        }
      },
      "NewsSummaryResponse": {
        "required": [
          "id",
          "title",
          "summary",
          "category",
          "publishedAt",
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultOfNewsSummaryResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Pagineringsconventie van de API (docs/05 §1): `?page=1&amp;pageSize=25`, maximaal 100."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PublicationRequest": {
        "required": [
          "visibility",
          "audienceRoles",
          "status",
          "publishAt"
        ],
        "type": "object",
        "properties": {
          "visibility": {
            "$ref": "#/components/schemas/ContentVisibility"
          },
          "audienceRoles": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "publishAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PublicationResponse": {
        "required": [
          "visibility",
          "audienceRoles",
          "status",
          "publishAt"
        ],
        "type": "object",
        "properties": {
          "visibility": {
            "$ref": "#/components/schemas/ContentVisibility"
          },
          "audienceRoles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "publishAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PublicationStatus": {
        "enum": [
          "Draft",
          "Scheduled",
          "Published",
          "Archived"
        ]
      },
...
    }
  },
...
}